Separate Category Selector to independent component
