+++ src/components/SettingsMenu.js
@@ -2,18 +2,11 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { MdClose } from 'react-icons/md';
 import { AiFillGithub, AiFillTwitterCircle } from 'react-icons/ai';
-import { useSelector, useDispatch } from 'react-redux';
-import { changeCategory } from '../redux/pageDetails/pageDetails';
 import TypeSelector from './TypeSelector';
+import CategorySelector from './CategorySelector';
 
 const SettingsMenu = (props) => {
-  const dispatch = useDispatch();
-  const { category } = useSelector((state) => state.pageDetails);
   const { setMenu } = props;
-
-  const onChangeHandlerCategory = (event) => {
-    dispatch(changeCategory(event.target.value));
-  };
 
   const closeSettings = (event) => {
     const { parentNode } = event.currentTarget;
@@ -30,23 +23,7 @@ const SettingsMenu = (props) => {
       </button>
       <form>
         <TypeSelector setMenu={setMenu} label="Anime Type" />
-        <label htmlFor="category">
-          <p>Anime Category</p>
-          <select onChange={onChangeHandlerCategory} name="category" id="category" value={category}>
-            <option value="All">All</option>
-            <option value="Action">Action</option>
-            <option value="Adventure">Adventure</option>
-            <option value="Comedy">Comedy</option>
-            <option value="Drama">Drama</option>
-            <option value="Fantasy">Fantasy</option>
-            <option value="Mystery">Mystery</option>
-            <option value="Sci-Fi">Sci-Fi</option>
-            <option value="Suspense">Suspense</option>
-            <option value="Romance">Romance</option>
-            <option value="Slice of Life">Slice of Life</option>
-            <option value="Supernatural">Supernatural</option>
-          </select>
-        </label>
+        <CategorySelector label="Anime Category" />
       </form>
       <div className="signature">
         <p> Made by JohnFTitor</p>
